+++ middleware/middleware_test.go
@@ -315,6 +315,36 @@ func TestMiddlewareServer(t *testing.T) {
 		attest.Equal(t, string(rb), postMsg)
 	})
 
+	t.Run("acme succeds", func(t *testing.T) {
+		t.Parallel()
+
+		msg := "hello world"
+		domain := "localhost"
+		o := WithLetsEncryptOpts(domain, "hey@example.com", 443, getSecretKey(), DirectIpStrategy, l)
+		wrappedHandler := All(someMiddlewareTestHandler(msg), o)
+
+		// Should not be a `NewTLSServer` since acme requires HTTP(not HTTPS)
+		ts := httptest.NewServer(
+			wrappedHandler,
+		)
+		defer ts.Close()
+
+		acmeChallengeURL := ts.URL + acmeURI
+		acmeChallengeURL = strings.ReplaceAll(acmeChallengeURL, "127.0.0.1", domain)
+		res, err := client.Get(acmeChallengeURL)
+		attest.Ok(t, err)
+
+		rb, err := io.ReadAll(res.Body)
+		attest.Ok(t, err)
+		defer res.Body.Close()
+
+		attest.Equal(t, res.StatusCode,
+			// Fails because the acmeHandler will get called with a host like `localhost:38355`
+			// which is not configured in `HostWhitelist`
+			http.StatusForbidden)
+		attest.Subsequence(t, string(rb), "not configured in HostWhitelist")
+	})
+
 	t.Run("concurrency safe", func(t *testing.T) {
 		t.Parallel()
 
@@ -442,10 +472,10 @@ func TestNew(t *testing.T) {
 			t.Parallel()
 			if tt.expectPanic {
 				attest.Panics(t, func() {
-					New(tt.domain, 443, nil, nil, nil, "secretKey", DirectIpStrategy, slog.Default())
+					New(tt.domain, "hey@example.com", letsEncryptStagingUrl, 443, nil, nil, nil, "secretKey", DirectIpStrategy, slog.Default())
 				})
 			} else {
-				New(tt.domain, 443, nil, nil, nil, "secretKey", DirectIpStrategy, slog.Default())
+				New(tt.domain, "hey@example.com", letsEncryptStagingUrl, 443, nil, nil, nil, "secretKey", DirectIpStrategy, slog.Default())
 			}
 		})
 	}