+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Most recent version is listed first.
 
 # v0.0.56
 - Set appropriate log level for http.Server.ErrorLog: https://github.com/komuw/ong/pull/288
+- Move acme handler to ong/middleware: https://github.com/komuw/ong/pull/290
 
 # v0.0.55
 - Improve timeouts: https://github.com/komuw/ong/pull/286